Refactoring and corrections

+++ src/AppBundle/Controller/REST/AccountSheetController.php
@@ -1,7 +1,7 @@
 <?php
 namespace AppBundle\Controller\REST;
 
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use AppBundle\Controller\REST\BaseRestController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
@@ -9,7 +9,7 @@ use FOS\RestBundle\Controller\Annotations as Rest;
 use AppBundle\Form\Type\AccountSheetType;
 use AppBundle\Entity\AccountSheet;
 
-class AccountSheetController extends Controller{
+class AccountSheetController extends BaseRestController{
     /**
      * @Rest\View()
      * @Rest\Get("/account/{account_id}/sheets")
@@ -19,9 +19,8 @@ class AccountSheetController extends Controller{
             ->getRepository('AppBundle:Account')
             ->find($account_id);
 
-        if (empty($account)) {
+        if(empty($account))
             return $this->notFound('Account');
-        }
 
         return $account->getAccountSheets();
     }
@@ -30,15 +29,15 @@ class AccountSheetController extends Controller{
      * @Rest\View()
      * @Rest\Get("/account/{account_id}/sheet/{sheet_id}")
      */
-    public function getAccountSheetAction($id, $sheet_id, Request $request){
+    public function getAccountSheetAction($account_id, $sheet_id, Request $request){
         $account = $this->get('doctrine.orm.entity_manager')
-                 ->getRepository('AppBundle:AccountSheet')
-                 ->find($id);
+                 ->getRepository('AppBundle:Account')
+                 ->find($account_id);
 
         if(empty($account))
-            return $this->notFound('AccountSheet');
+            return $this->notFound('Account');
 
-        return $account->getOperations();
+        return $account->getSheetById($sheet_id);
     }
 
     /**
@@ -51,7 +50,7 @@ class AccountSheetController extends Controller{
                  ->find($account_id);
 
         if(empty($account))
-            return $this->notFound('AccountSheet');
+            return $this->notFound('Account');
 
         $aSheet = new AccountSheet();
         $aSheet->setAccount($account);
@@ -72,24 +71,26 @@ class AccountSheetController extends Controller{
 
     /**
      * @Rest\View()
-     * @Rest\Put("/sheet/{id}")
+     * @Rest\Put("/account/{account_id}/sheet/{sheet_id}")
      */
-    public function putAccountSheetAction($id, Request $request){
-        $aSheet = $this->get('doctrine.orm.entity_manager')
-                    ->getRepository('AppBundle:AccountSheet')
-                    ->find($id);
+    public function putAccountSheetAction($account_id, $sheet_id, Request $request){
+        $em = $this->get('doctrine.orm.entity_manager');
+        $account = $em->getRepository('AppBundle:Account')
+                    ->find($account_id);
 
-        if (empty($aSheet)) {
-            return new JsonResponse(['message' => 'Sheet not found'], Response::HTTP_NOT_FOUND);
-        }
+        if(empty($account))
+            return $this->notFound('Account');
+
+        $aSheet = $account->getSheetById($sheet_id);
+
+        if(empty($aSheet))
+            return $this->notFound('AccountSheet');
 
         $form = $this->createForm(AccountSheetType::class, $aSheet);
 
         $form->submit($request->request->all());
 
         if ($form->isValid()) {
-            $em = $this->get('doctrine.orm.entity_manager');
-
             $em->merge($aSheet);    // Pas nécessaire, juste pas mesure de clarté
             $em->flush();
             return $aSheet;
@@ -100,19 +101,20 @@ class AccountSheetController extends Controller{
 
     /**
      * @Rest\View(statusCode=Response::HTTP_NO_CONTENT)
-     * @Rest\Delete("/account/sheet/{id}")
+     * @Rest\Delete("/account/{account_id}/sheet/{sheet_id}")
      */
-    public function removeAccountSheetAction($id, Request $request){
+    public function removeAccountSheetAction($account_id, $sheet_id, Request $request){
         $em = $this->get('doctrine.orm.entity_manager');
-        $aSheet = $em->getRepository('AppBundle:AccountSheet')
-                    ->find($id);
+        $account = $em->getRepository('AppBundle:Account')->find($account_id);
+
+        if(empty($account))
+            return $this->notFound('Account');
+
+        $aSheet = $account->getSheetById($sheet_id);
+
         if($aSheet){
             $em->remove($aSheet);
             $em->flush();
         }
-    }
-
-    private function notFound($type){
-        return \FOS\RestBundle\View\View::create(['message' => "$type not found"], Response::HTTP_NOT_FOUND);
     }
 }
